moved add answer form into its own component
